Move blob helpers to right file
