+++ tests/core/pyspec/eth2spec/test/deneb/fork_choice/test_on_block.py
@@ -4,19 +4,10 @@ from eth2spec.test.context import (
     spec_state_test,
     with_all_phases_from_except,
 )
-from eth2spec.test.helpers.blob import (
-    get_sample_blob_tx,
-)
-from eth2spec.test.helpers.block import (
-    build_empty_block_for_next_slot,
-)
 from eth2spec.test.helpers.constants import (
     DENEB,
     EIP7732,
     FULU,
-)
-from eth2spec.test.helpers.execution_payload import (
-    compute_el_block_hash,
 )
 from eth2spec.test.helpers.fork_choice import (
     BlobData,
@@ -27,32 +18,7 @@ from eth2spec.test.helpers.fork_choice import (
 from eth2spec.test.helpers.state import (
     state_transition_and_sign_block,
 )
-
-
-def get_block_with_blob(spec, state, rng=None, blob_count=1):
-    block = build_empty_block_for_next_slot(spec, state)
-    opaque_tx, blobs, blob_kzg_commitments, blob_kzg_proofs = get_sample_blob_tx(
-        spec, blob_count=blob_count, rng=rng
-    )
-    block.body.execution_payload.transactions = [opaque_tx]
-    block.body.execution_payload.block_hash = compute_el_block_hash(
-        spec, block.body.execution_payload, state
-    )
-    block.body.blob_kzg_commitments = blob_kzg_commitments
-    return block, blobs, blob_kzg_proofs
-
-
-def get_block_with_blob_and_sidecars(spec, state, rng=None, blob_count=1):
-    state = state.copy()
-
-    block, blobs, blob_kzg_proofs = get_block_with_blob(spec, state, rng=rng, blob_count=blob_count)
-    cells_and_kzg_proofs = [spec.compute_cells_and_kzg_proofs(blob) for blob in blobs]
-
-    # We need a signed block to call `get_data_column_sidecars_from_block`
-    signed_block = state_transition_and_sign_block(spec, state, block)
-
-    sidecars = spec.get_data_column_sidecars_from_block(signed_block, cells_and_kzg_proofs)
-    return block, blobs, blob_kzg_proofs, signed_block, sidecars
+from eth2spec.test.helpers.blob import get_block_with_blob
 
 
 # TODO(jtraglia): Use with_all_phases_from_to_except after EIP7732 is based on Fulu.
